pipecat-backend/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
cartesia==2.0.6
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.2
coloredlogs==15.0.1
coverage==7.10.6
cryptography==45.0.6
daily-python==0.21.0
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.19.1
firebase-admin==6.5.0
flatbuffers==25.2.10
frozenlist==1.7.0
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-speech==2.32.0
google-cloud-storage==3.3.0
google-cloud-texttospeech==2.26.0
google-crc32c==1.7.1
google-genai==1.24.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
helpers==0.2.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy==1.11.2
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.20.1
openai==1.98.0
orjson==3.11.1
packaging==25.0
pillow==11.3.0
pipecat-ai==0.0.92
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.0.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyloudnorm==0.1.1
pyparsing==3.2.3
pypdf==5.1.0
pytest==8.3.3
pytest-asyncio==0.25.3
pytest-cov==5.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
resampy==0.4.3
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
ruff==0.6.9
scipy==1.16.1
sentry-sdk==2.34.1
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
soxr==0.5.0.post1
starlette==0.47.2
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.11.0
tqdm==4.67.1
typer==0.16.0
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==14.2
wheel==0.45.1
yarl==1.20.1
